+++ instagram_private_api/__init__.py
@@ -3,4 +3,4 @@ from .compatpatch import ClientCompatPatch
 from .errors import ClientError, ClientLoginError, ClientLoginRequiredError, ClientCookieExpiredError
 
 
-__version__ = '1.0.5'
+__version__ = '1.0.6'